\name{ORBayesTtest.app}
\alias{ORBayesTtestApp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A Objective and Robust Bayesian $t$-test
}
\description{
Run shiny app to compute the Bayes Factor based on the intrinsic, Jeffrey's and Berger's robust prior. For a population mean
}
\usage{
ORBayesianTtest.app()
}
\value{
Run Shiny App
}
\references{
Almodóvar-Rivera, I. A., & Pericchi-Guerra, L. R. (2024). An objective and robust bayes factor for the hypothesis test one sample and two population means. Entropy, 26(1), 88.
}
\author{
Israel A. Almodóvar-Rivera, Ph.D.
}
\examples{
\dontrun{
ORBayesianTtest.app()
}
}